chore(test): move reporting tests to its own thing for growth
